Copy into ./man/files the figures required to support lifecycle badges.

copy_lifecycle_figures(new_path = here::here("man", "figures"))

Arguments

new_path
Destination path of the copied figures.

Value

Invisible new_path.

Details

To support lifecycle badges you need to first run use_template("compat-lifecycle.R", package = "fgeo.template"). See also https://github.com/r-lib/rlang/blob/master/R/compat-lifecycle.R.

Examples

# NOT RUN {
# This creates a file in the root of your package. Move it to R/
use_template("compat-lifecycle.R", package = "fgeo.template")
# This adds the images for the badges into man/figures
copy_lifecycle_figures()
# }